+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # @commercetools-frontend/b2b-launchpad
 
+## 1.1.0
+
+### Minor Changes
+
+-   e27d142: Add method to remove and add recurrence info from line items
+
+### Patch Changes
+
+-   50899cf: Fix logs and bump nanoid package to fix security issue
+-   1977d61: Streamline recurrence policy handling in CartMapper
+
 ## 1.0.0
 
 ### Major Changes